+++ Sources/StreamChat/Query/Sorting/Sorting.swift
@@ -43,3 +43,76 @@ public struct Sorting<Key: SortingKey>: Encodable, CustomStringConvertible {
 
 extension Sorting: Equatable where Key: Equatable {}
 extension Sorting: Hashable where Key: Hashable {}
+
+/// A sorting key that can be converted to local DB query or local runtime sorting.
+public struct LocalConvertibleSortingKey<Model>: SortingKey, Encodable, Equatable {
+    let keyPath: PartialKeyPath<Model>?
+    let localKey: String?
+    let remoteKey: String
+    var requiresRuntimeSorting: Bool {
+        localKey == nil
+    }
+
+    init(keyPath: PartialKeyPath<Model>?, localKey: String?, remoteKey: String) {
+        self.keyPath = keyPath
+        self.localKey = localKey
+        self.remoteKey = remoteKey
+    }
+
+    init(localKey: String?, remoteKey: String) {
+        keyPath = nil
+        self.localKey = localKey
+        self.remoteKey = remoteKey
+    }
+
+    public static func custom<Value>(keyPath: KeyPath<Model, Value>, key: String) -> Self {
+        .init(keyPath: keyPath, localKey: nil, remoteKey: key)
+    }
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        try container.encode(remoteKey)
+    }
+}
+
+extension LocalConvertibleSortingKey: CustomDebugStringConvertible {
+    public var debugDescription: String {
+        remoteKey
+    }
+
+    func sortDescriptor(isAscending: Bool) -> NSSortDescriptor? {
+        guard let localKey = self.localKey else {
+            return nil
+        }
+        return .init(key: localKey, ascending: isAscending)
+    }
+
+    func sortValue(isAscending: Bool) -> SortValue<Model>? {
+        guard let keyPath = keyPath else {
+            return nil
+        }
+        return SortValue(keyPath: keyPath, isAscending: isAscending)
+    }
+}
+
+/// A protocol for queries that can be converted to local sorting.
+protocol LocalConvertibleSortingQuery {
+    associatedtype Model
+    var sort: [Sorting<LocalConvertibleSortingKey<Model>>] { get }
+}
+
+extension LocalConvertibleSortingQuery {
+    var requiresRuntimeSorting: Bool {
+        sort.contains { $0.key.requiresRuntimeSorting }
+    }
+
+    /// Returns the sort values for runtime sorting.
+    /// If one of the sort keys requires runtime sorting, all sort values will be used
+    /// for runtime sorting.
+    var runtimeSortingValues: [SortValue<Model>] {
+        if !requiresRuntimeSorting {
+            return []
+        }
+        return sort.compactMap { $0.key.sortValue(isAscending: $0.isAscending) }
+    }
+}